Various GUI and Color related fixes/improvements. UILog now actually works. UIList now only renders those elements which are relevant (can have huge lists with less lag). Noticed problem with Random.Randi() function, Randf() works better

+++ src/UI/UIImage.cpp
@@ -87,7 +87,7 @@ void UIImage::RenderSelf(GraphicsState & graphicsState)
 		///
 		if (shader->uniformModelMatrix != -1){
 			/// TRanslatem power !
-			Matrix4d * model = &graphicsState.modelMatrixD;
+			Matrix4f * model = &graphicsState.modelMatrixD;
 			float transX = alignmentX * parent->sizeX;
 			float transY = alignmentY * parent->sizeY;
 			model->Translate(transX,transY,0);
